+++ lib/shared/src/codebase-context/messages.ts
@@ -37,6 +37,7 @@ export interface PreciseContext {
 
 export interface HoverContext {
     symbolName: string
+    sourceSymbolName?: string
     type: 'definition' | 'typeDefinition' | 'implementation'
     content: string[]