+++ vscode/scripts/download-wasm-modules.ts
@@ -4,7 +4,7 @@ import path from 'path'
 
 import ProgressBar from 'progress'
 
-import { ROOT_PATH } from '@sourcegraph/cody-shared/src/common'
+import { ROOT_PATH } from '@sourcegraph/cody-shared/src/common/paths'
 
 const DIST_DIRECTORY = path.join(ROOT_PATH, 'vscode/dist')
 const WASM_DIRECTORY = path.join(ROOT_PATH, 'vscode/resources/wasm')
@@ -92,7 +92,7 @@ function getMissingFiles(urls: string[]): string[] {
 
 function getFilePathFromURL(url: string): string {
     const parts = url.split('/')
-    return parts[parts.length - 1]
+    return parts.at(-1)!
 }
 
 function downloadFile(url: string): Promise<WriteStream> {