+++ vscode/src/completions/context/graph-section-observer.ts
@@ -208,15 +208,15 @@ export class GraphSectionObserver implements vscode.Disposable, GraphContextFetc
             }
         }
 
-        if (!section.preloadedContext) {
+        if (section.preloadedContext) {
+            section.preloadedContext.lastRevalidateAt = Date.now()
+            section.preloadedContext.isStale = false
+        } else {
             section.preloadedContext = {
                 lastRevalidateAt: Date.now(),
                 graphContext: null,
                 isStale: false,
             }
-        } else {
-            section.preloadedContext.lastRevalidateAt = Date.now()
-            section.preloadedContext.isStale = false
         }
 
         debugSubscriber.notify()
@@ -248,7 +248,7 @@ export class GraphSectionObserver implements vscode.Disposable, GraphContextFetc
                 const isSelected =
                     selectedDocument?.uri.toString() === document.uri.toString() &&
                     selections?.some(selection => section.location.range.contains(selection))
-                const isLast = document.sections[document.sections.length - 1] === section
+                const isLast = document.sections.at(-1) === section
                 const isHydrated = !!section.preloadedContext
                 const isStale = section.preloadedContext?.isStale ?? false
 
@@ -455,10 +455,15 @@ function hoverContextsToSnippets(contexts: HoverContext[]): SymbolContextSnippet
 }
 
 function hoverContextToSnippets(context: HoverContext): SymbolContextSnippet {
-    console.log({ context })
+    const sourceSymbolAndRelationship =
+        context.sourceSymbolName && context.type !== 'definition'
+            ? { symbol: context.sourceSymbolName, relationship: context.type }
+            : undefined
+
     return {
         fileName: path.normalize(vscode.workspace.asRelativePath(URI.parse(context.uri).fsPath)),
         symbol: context.symbolName,
+        sourceSymbolAndRelationship,
         content: context.content.join('\n').trim(),
     }
 }