+++ vscode/src/completions/providers/anthropic.ts
@@ -12,6 +12,7 @@ import {
     CLOSING_CODE_TAG,
     extractFromCodeBlock,
     fixBadCompletionStart,
+    formatSymbolContextRelationship,
     getHeadAndTail,
     MULTILINE_STOP_SEQUENCE,
     OPENING_CODE_TAG,
@@ -24,6 +25,8 @@ import { forkSignal, messagesToText } from '../utils'
 import { CompletionProviderTracer, Provider, ProviderConfig, ProviderOptions } from './provider'
 
 const CHARS_PER_TOKEN = 4
+export const MULTI_LINE_STOP_SEQUENCES = [anthropic.HUMAN_PROMPT, CLOSING_CODE_TAG]
+export const SINGLE_LINE_STOP_SEQUENCES = [anthropic.HUMAN_PROMPT, CLOSING_CODE_TAG, MULTILINE_STOP_SEQUENCE]
 
 function tokensToChars(tokens: number): number {
     return tokens * CHARS_PER_TOKEN
@@ -103,13 +106,15 @@ export class AnthropicProvider extends Provider {
 
         const { head, tail, overlap } = getHeadAndTail(this.options.docContext.prefix)
 
+        // Infill block represents the code we want the model to complete
+        const infillBlock = `${tail.trimmed.trimEnd()}`
         // code before the cursor, after removing the code for the infillBlock
         // Using this instead of head.trimmed to preserve the spacing from prefix so the model can determines the patterns of surrounding code
-        const infillPrefix = this.options.docContext.prefix.replace(tail.trimmed.trim(), '')
+        // Use regex to makes sure only the last trimmedTail match is replaced to avoid replacing overlapping code
+        const infillBlockRegex = new RegExp(`${infillBlock}\\s*$`, 'g')
+        const infillPrefix = this.options.docContext.prefix.replace(infillBlockRegex, '')
         // code after the cursor
         const infillSuffix = this.options.docContext.suffix
-        // Infill block represents the code we want the model to complete
-        const infillBlock = `${tail.trimmed}`
 
         const prefixMessagesWithInfill: Message[] = [
             {
@@ -118,11 +123,11 @@ export class AnthropicProvider extends Provider {
             },
             {
                 speaker: 'assistant',
-                text: 'I am a code completion AI with exceptional context-awareness designed to auto-complete nested code blocks with high-quality code that seamlessly integrates with surrounding code without duplicating existing implementations.',
+                text: 'I am a code completion AI with exceptional context-awareness designed to auto-complete nested code blocks with high-quality code that seamlessly integrates with surrounding code.',
             },
             {
                 speaker: 'human',
-                text: `Below is the code from file path ${this.options.fileName}. Detect the functionality, formats, style, patterns, and logics in use from code outside ${OPENING_CODE_TAG} XML tags. Then, use what you detect and reuse assetmethods/libraries to complete and enclose complete code only inside ${OPENING_CODE_TAG} XML tags precisely:
+                text: `Below is the code from file path ${this.options.fileName}. Detect the functionality, formats, style, patterns, and logics in use from code outside ${OPENING_CODE_TAG} XML tags. Then, use what you detect and reuse assetmethods/libraries to complete and enclose completed code only inside ${OPENING_CODE_TAG} tags precisely without duplicating existing implementations. Here is the code:
                 ${infillPrefix}${OPENING_CODE_TAG}${CLOSING_CODE_TAG}${infillSuffix}`,
             },
             {
@@ -151,7 +156,9 @@ export class AnthropicProvider extends Provider {
                     speaker: 'human',
                     text:
                         'symbol' in snippet && snippet.symbol !== ''
-                            ? `Additional documentation for \`${snippet.symbol}\`: ${OPENING_CODE_TAG}${snippet.content}${CLOSING_CODE_TAG}`
+                            ? `Additional documentation for \`${snippet.symbol}\`${formatSymbolContextRelationship(
+                                  snippet.sourceSymbolAndRelationship
+                              )}: ${OPENING_CODE_TAG}${snippet.content}${CLOSING_CODE_TAG}`
                             : `Codebase context from file path '${snippet.fileName}': ${OPENING_CODE_TAG}${snippet.content}${CLOSING_CODE_TAG}`,
                 },
                 {
@@ -186,13 +193,13 @@ export class AnthropicProvider extends Provider {
                   temperature: 0.5,
                   messages: prompt,
                   maxTokensToSample: this.responseTokens,
-                  stopSequences: [anthropic.HUMAN_PROMPT, CLOSING_CODE_TAG],
+                  stopSequences: MULTI_LINE_STOP_SEQUENCES,
               }
             : {
                   temperature: 0.5,
                   messages: prompt,
                   maxTokensToSample: Math.min(50, this.responseTokens),
-                  stopSequences: [anthropic.HUMAN_PROMPT, CLOSING_CODE_TAG, MULTILINE_STOP_SEQUENCE],
+                  stopSequences: SINGLE_LINE_STOP_SEQUENCES,
               }
         tracer?.params(args)